Remove `RemoveSuperTypeByPackage` as it's likely to make unchecked changes
